+++ src/graphics/gfx_print.cpp
@@ -55,7 +55,7 @@ int SuperTextPixLen(int SuperN, const char* SuperChars, int Font)
     return FontManager::textSize(SuperChars, SuperN, dFont, FontManager::fontSizeFromSmbxFont(Font)).w();
 }
 
-void SuperPrintRightAlign(int SuperN, const char* SuperChars, int Font, float X, float Y, XTColor color)
+void SuperPrintRightAlign(int SuperN, const char* SuperChars, int Font, int X, int Y, XTColor color)
 {
     int RealFont = Font;
     bool outline = false;
@@ -80,7 +80,7 @@ void SuperPrintRightAlign(int SuperN, const char* SuperChars, int Font, float X,
     SuperPrint(SuperN, SuperChars, RealFont, X, Y, color);
 }
 
-void SuperPrintCenter(int SuperN, const char* SuperChars, int Font, float X, float Y, XTColor color)
+void SuperPrintCenter(int SuperN, const char* SuperChars, int Font, int X, int Y, XTColor color)
 {
     int RealFont = Font;
     bool outline = false;
@@ -105,7 +105,7 @@ void SuperPrintCenter(int SuperN, const char* SuperChars, int Font, float X, flo
     SuperPrint(SuperN, SuperChars, RealFont, X, Y, color);
 }
 
-void SuperPrintScreenCenter(int SuperN, const char* SuperChars, int Font, float Y, XTColor color)
+void SuperPrintScreenCenter(int SuperN, const char* SuperChars, int Font, int Y, XTColor color)
 {
     int RealFont = Font;
     bool outline = false;
@@ -121,16 +121,16 @@ void SuperPrintScreenCenter(int SuperN, const char* SuperChars, int Font, float 
 
     if(dFont >= 0)
     {
-        float X = (XRender::TargetW / 2) - (FontManager::textSize(SuperChars, SuperN, dFont, FontManager::fontSizeFromSmbxFont(Font)).w() / 2);
+        int X = (XRender::TargetW / 2) - (FontManager::textSize(SuperChars, SuperN, dFont, FontManager::fontSizeFromSmbxFont(Font)).w() / 2);
         FontManager::printText(SuperChars, SuperN, X, Y, dFont, color, FontManager::fontSizeFromSmbxFont(Font), outline);
         return;
     }
 
-    float X = (XRender::TargetW / 2) - (SuperTextPixLen(SuperN, SuperChars, RealFont) / 2);
+    int X = (XRender::TargetW / 2) - (SuperTextPixLen(SuperN, SuperChars, RealFont) / 2);
     SuperPrint(SuperN, SuperChars, RealFont, X, Y, color);
 }
 
-void SuperPrint(int SuperN, const char* SuperChars, int Font, float X, float Y,
+void SuperPrint(int SuperN, const char* SuperChars, int Font, int X, int Y,
                 XTColor color)
 {
     bool outline = false;
@@ -161,25 +161,25 @@ int SuperTextPixLen(const char* SuperChars, int Font)
     return SuperTextPixLen(len, SuperChars, Font);
 }
 
-void SuperPrintRightAlign(const char* SuperChars, int Font, float X, float Y, XTColor color)
+void SuperPrintRightAlign(const char* SuperChars, int Font, int X, int Y, XTColor color)
 {
     int len = (int)SDL_strlen(SuperChars);
     SuperPrintRightAlign(len, SuperChars, Font, X, Y, color);
 }
 
-void SuperPrintCenter(const char* SuperChars, int Font, float X, float Y, XTColor color)
+void SuperPrintCenter(const char* SuperChars, int Font, int X, int Y, XTColor color)
 {
     int len = (int)SDL_strlen(SuperChars);
     SuperPrintCenter(len, SuperChars, Font, X, Y, color);
 }
 
-void SuperPrintScreenCenter(const char* SuperChars, int Font, float Y, XTColor color)
+void SuperPrintScreenCenter(const char* SuperChars, int Font, int Y, XTColor color)
 {
     int len = (int)SDL_strlen(SuperChars);
     SuperPrintScreenCenter(len, SuperChars, Font, Y, color);
 }
 
-void SuperPrint(const char* SuperChars, int Font, float X, float Y, XTColor color)
+void SuperPrint(const char* SuperChars, int Font, int X, int Y, XTColor color)
 {
     int len = (int)SDL_strlen(SuperChars);
     SuperPrint(len, SuperChars, Font, X, Y, color);
@@ -193,22 +193,22 @@ int SuperTextPixLen(const std::string& SuperWords, int Font)
     return SuperTextPixLen((int)SuperWords.size(), SuperWords.c_str(), Font);
 }
 
-void SuperPrintRightAlign(const std::string& SuperWords, int Font, float X, float Y, XTColor color)
+void SuperPrintRightAlign(const std::string& SuperWords, int Font, int X, int Y, XTColor color)
 {
     SuperPrintRightAlign((int)SuperWords.size(), SuperWords.c_str(), Font, X, Y, color);
 }
 
-void SuperPrintCenter(const std::string& SuperWords, int Font, float X, float Y, XTColor color)
+void SuperPrintCenter(const std::string& SuperWords, int Font, int X, int Y, XTColor color)
 {
     SuperPrintCenter((int)SuperWords.size(), SuperWords.c_str(), Font, X, Y, color);
 }
 
-void SuperPrintScreenCenter(const std::string& SuperWords, int Font, float Y, XTColor color)
+void SuperPrintScreenCenter(const std::string& SuperWords, int Font, int Y, XTColor color)
 {
     SuperPrintScreenCenter((int)SuperWords.size(), SuperWords.c_str(), Font, Y, color);
 }
 
-void SuperPrint(const std::string& SuperWords, int Font, float X, float Y, XTColor color)
+void SuperPrint(const std::string& SuperWords, int Font, int X, int Y, XTColor color)
 {
     SuperPrint((int)SuperWords.size(), SuperWords.c_str(), Font, X, Y, color);
 }
